+++ src/main/java/com/loschiferos/ztech/profile/interfaces/rest/resources/CreateSupplierResource.java
@@ -5,7 +5,7 @@ public record CreateSupplierResource(
         String email,
         String password,
         String address,
-        int phone,
+        Long phone,
         String photo,
         String ruc
 ) {
