+++ src/main/java/com/loschiferos/ztech/profile/interfaces/rest/resources/PlantOwnerResource.java
@@ -8,9 +8,9 @@ public record PlantOwnerResource(
         String email,
         String password,
         String address,
-        int phone,
+        Long phone,
         String photo,
-        int dni,
+        Long dni,
         Date birthday,
         String gender
 ) {
